requirements.txt
aiohttp
asyncache~=0.3.1
Babel~=2.15.0
beautifulsoup4~=4.12.0
cryptography~=42.0.0
discord.py~=2.3.0
emoji~=2.11.0
feedparser~=6.0.0
geocoder
GitPython
google-api-python-client~=2.129.0
i18nice~=0.15.0
imageio
isbnlib
LRFutils==0.1.2
mysql-connector-python>=8.0
nltk>=3.7
Pillow~=10.3.0
psutil>=5.8
python-dateutil
pytz
rembg~=2.0.0
timezonefinder

# Antiscam training
pandas
scikit-learn
